helpers: pmaports: Always fall back to parsing all APKBUILDS
Previously, it would only fall back to parsing all APKBUILDs if a main
package was guessed for this potential subpackage. We do however want to
scan all APKBUILDs in case the package is provided by one of the
packages in pmaports. Also, remove the enforcement of = in provided
packages, since the postmarketOS initramfs does not respect this
anymore.

Signed-off-by: Jens Reidel <adrian@travitia.xyz>